+++ src/Locations/data/LocationNames.ts
@@ -24,6 +24,7 @@ export enum LocationName {
     AevumSnapFitnessGym             = "Snap Fitness Gym",
     AevumSummitUniversity           = "Summit University",
     AevumWatchdogSecurity           = "Watchdog Security",
+    AevumCasino                     = "Iker Molina Casino",
 
     // Chongqing locations
     ChongqingKuaiGongInternational  = "KuaiGong International",