+++ src/PersonObjects/IPlayer.ts
@@ -179,4 +179,5 @@ export interface IPlayer {
     giveExploit(exploit: Exploit): void;
     queryStatFromString(str: string): number;
     getIntelligenceBonus(weight: number): number;
+    getCasinoWinnings(): number;
 }