+++ src/ui/CharacterInfo.tsx
@@ -42,7 +42,7 @@ export function CharacterInfo(p: IPlayer): React.ReactElement {
         if(!(p.bitNodeN === 9 || SourceFileFlags[9] > 0)) {
             return <><span>{`Hacknet Nodes owned: ${p.hacknetNodes.length}</span>`}</span><br /></>
         } else {
-            return <><span>{`Hacknet Servers owned: ${p.hacknetNodes.length} / ${HacknetServerConstants.MaxServers}</span>`}</span><br /></>
+            return <><span>{`Hacknet Servers owned: ${p.hacknetNodes.length} / ${HacknetServerConstants.MaxServers}`}</span><br /></>
         }
     }
 
@@ -59,6 +59,7 @@ export function CharacterInfo(p: IPlayer): React.ReactElement {
         if (src.hospitalization) { parts.push([`Hospitalization:`, Money(src.hospitalization)]) };
         if (src.infiltration)    { parts.push([`Infiltration:`, Money(src.infiltration)]) };
         if (src.stock)           { parts.push([`Stock Market:`, Money(src.stock)]) };
+        if (src.casino)          { parts.push([`Casino:`, Money(src.casino)]) };
 
         return StatsTable(parts, "");
     }