+++ doc/source/changelog.rst
@@ -3,22 +3,40 @@
 Changelog
 =========
 
+v0.51.2 - 2021-04-09 Vegas, Baby! (hydroflame)
+----------------------------------------------
+
+**New location: The Iker Molina Casino**
+
+* A casino opened in Aevum. However the house is rumored to cheat. If only 
+  we could give them a taste of their own medicine.
+
+**Misc.**
+
+* Link to discord added under options
+* 'getMemberInformation' doc updated, oops
+* tech vendor now handle max ram and cores.
+
 v0.51.1 - 2021-04-06 Bugfixes because the author of the last patch sucks (it's hydroflame)
 ------------------------------------------------------------------------------------------
 
 **Netscript**
+
 * 'getPlayer' returns players faction and tor
 * 'hospitalization' is a new singularity function.
 * 'gang.getMemberInformation' now returns more information.
 * 'hacknet.hashCapacity' is a new hacknet function that returns the maximum hash capacity.
 
 **Hospitalization**
+
 * Now only cost at most 10% of your money.
 
 **Bugfix**
+
 * confirmation dialog box no longer use previous text
 
 **Accessibility**
+
 * The game is a little easier to handle for screen readers (yes, there's an
   absolute legend playing this game with a screen reader)
 * Infiltration use buttons instead of a-links
@@ -26,6 +44,7 @@ v0.51.1 - 2021-04-06 Bugfixes because the author of the last patch sucks (it's h
   map display as a list of buttons.
 
 **Misc.**
+
 * 'fl1ght.exe' will no longer suggest the combat path. Related faction
   requirements unchanged.